requirements.txt
pandas>=2.2
numpy>=2.1
scikit-learn>=1.5
fastapi
uvicorn
pydantic
requests
google-generativeai
python-dotenv
